centralize hover intent
