+++ web-common/src/components/tooltip/Tooltip.svelte
@@ -17,7 +17,6 @@ You can set the following parameters:
 - hoverIntentTimeout: the time in milliseconds to wait for hover intent.
 - active: a boolean to control the active state of the tooltip.
 
-
 The suppress prop gives programmatic access to suppressing the tooltip, e.g. if a menu is open or you just need
 to stop all tooltips.
 
@@ -34,7 +33,7 @@ FIXME: In the future, we should also be listening to focus events from the child
     Alignment,
     Location,
   } from "@rilldata/web-common/lib/place-element";
-  import { onDestroy } from "svelte";
+  import HoverIntent from "./HoverIntent.svelte";
 
   export let location: Location = "bottom";
   export let alignment: Alignment = "middle";
@@ -48,35 +47,6 @@ FIXME: In the future, we should also be listening to focus events from the child
   export let active = false;
 
   let parent: HTMLDivElement;
-  let waitUntilTimer: ReturnType<typeof setTimeout> | undefined;
-  let hoverIntentTimer: ReturnType<typeof setTimeout> | undefined;
-  let resetMoveTimer: ReturnType<typeof setTimeout> | undefined;
-  let lastMouseX = 0;
-  let lastMouseY = 0;
-  let mouseMoved = false;
-  let isHovering = false;
-
-  function clearAllTimers() {
-    if (waitUntilTimer) {
-      clearTimeout(waitUntilTimer);
-      waitUntilTimer = undefined;
-    }
-    if (hoverIntentTimer) {
-      clearTimeout(hoverIntentTimer);
-      hoverIntentTimer = undefined;
-    }
-    if (resetMoveTimer) {
-      clearTimeout(resetMoveTimer);
-      resetMoveTimer = undefined;
-    }
-  }
-
-  function waitUntil(callback, time = activeDelay) {
-    clearAllTimers();
-    waitUntilTimer = setTimeout(() => {
-      callback();
-    }, time);
-  }
 
   /** create child-supported suppression.
    * If a child changes this context store to true, we should
@@ -90,78 +60,23 @@ FIXME: In the future, we should also be listening to focus events from the child
     childRequestedTooltipSuppression,
   );
 
-  function handlePointerEnter(event: PointerEvent) {
-    isHovering = true;
-    lastMouseX = event.clientX;
-    lastMouseY = event.clientY;
-    mouseMoved = false;
-
-    clearAllTimers();
-
-    hoverIntentTimer = setTimeout(() => {
-      if (!mouseMoved && isHovering) {
-        waitUntil(() => {
-          active = true;
-        });
-      }
-    }, hoverIntentTimeout);
+  function handleActive(event: CustomEvent<boolean>) {
+    active = event.detail;
   }
-
-  function resetMoveState() {
-    mouseMoved = false;
-    // Start a new hover intent check
-    if (isHovering) {
-      hoverIntentTimer = setTimeout(() => {
-        if (!mouseMoved && isHovering) {
-          waitUntil(() => {
-            active = true;
-          });
-        }
-      }, hoverIntentTimeout);
-    }
-  }
-
-  function handlePointerMove(event: PointerEvent) {
-    if (!isHovering) return;
-
-    const deltaX = Math.abs(event.clientX - lastMouseX);
-    const deltaY = Math.abs(event.clientY - lastMouseY);
-
-    if (deltaX > hoverIntentThreshold || deltaY > hoverIntentThreshold) {
-      mouseMoved = true;
-      clearAllTimers();
-
-      // Set a timer to reset the move state after movement stops
-      resetMoveTimer = setTimeout(resetMoveState, hoverIntentTimeout);
-    }
-
-    lastMouseX = event.clientX;
-    lastMouseY = event.clientY;
-  }
-
-  function handlePointerLeave() {
-    isHovering = false;
-    mouseMoved = false;
-    clearAllTimers();
-    waitUntil(() => {
-      active = false;
-    }, nonActiveDelay);
-  }
-
-  onDestroy(() => {
-    clearAllTimers();
-  });
 </script>
 
-<div
-  role="tooltip"
-  class="contents"
-  bind:this={parent}
-  on:pointerenter={handlePointerEnter}
-  on:pointermove={handlePointerMove}
-  on:pointerleave={handlePointerLeave}
->
-  <slot />
+<div class="contents">
+  <HoverIntent
+    threshold={hoverIntentThreshold}
+    timeout={hoverIntentTimeout}
+    {activeDelay}
+    {nonActiveDelay}
+    on:active={handleActive}
+  >
+    <div bind:this={parent}>
+      <slot />
+    </div>
+  </HoverIntent>
   {#if active && !suppress && !$childRequestedTooltipSuppression}
     <FloatingElement target={parent} {location} {alignment} {distance} {pad}>
       <slot name="tooltip-content" />
